+++ changelog.md
@@ -2,8 +2,9 @@ ChangeLog
 =========
 1.2.15 (2020-12-09)
 -------------------------
-* Changed GUI
+* Changes in GUI
 * Fix for import bookmark with quote
+* Fix for encoding
 
 
 1.2.14 (2020-12-05)